+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased](https://github.com/xsolla/xsolla-sdk-php/compare/v2.4.0...master)
 ### Fixed
 * [BC BREAK] Added required integer parameters `limit` and `offset` to `XsollaClient::ListEvents`.
+* [BC BREAK] Renamed Frontend API to [Storefront](http://developers.xsolla.com/api.html#storefront)
 
 ## [v2.4.0](https://github.com/xsolla/xsolla-sdk-php/compare/v2.3.0...v2.4.0) - 2016-04-27
 ### Added